langmaster_server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
openai-whisper==20231117
ffmpeg-python==0.2.0
edge-tts==6.1.9
python-multipart==0.0.6
